refine the logic for BatchCompactOperator
